src/Makefile: Add a target to build the headers
This adds a 'hdrchk' target that builds the header files, This is to
help ensure the header files remain self-contained, i.e that they have
everything that they need to be built individually.

This slightly changes how the dependencies work, making sure that the .c
and .h files, of which some have the same name, produce differently
named dependency files. So rather than just having the filename (minus
the suffix) .d files we now have filename.o.d for the C files and
filename.gch.d for the header files.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>