src/Makefile: Reduce some duplication
Rather than having rules for each of the source directories, just have a
rule for all the .c files and a rule for all the .h files.

As part of this we now put all the generated .o and .gch files into a
.objs directory.

Although this is perhaps not a bad idea anyway, it was kind of required
to make the rules like

  %.o: %.c $(DEPDIR)/%.o.d

work.

We store the source and object file lists as just the filenames without
directory prefixes and use vpath to allow make to find the source files.

We then have

  $(OBJ_DIR)/%.o: %.c $(DEPDIR)/%.o.d

otherwise the % would expand out to the full path of the file and that
would break the $(DEPDIR)/%.o.d part as we are just using a flat
directory structure there and make would always end up rebuilding the
platform/*/*.c files as it would be looking for e.g
.d/platform/linux/gen_uuid.o.d when in fact it is .d/gen_uuid.o.d and we
can't just fix the path in the rule as % is expanded out after all other
expansions/functions etc.

This also tweaks the make output slightly in that instead of showing e.g

  CC     auth.o

we now show

  CC     auth.c

seeing as the .o files are now all being stored under .objs and if we
just used $@ they'd all show up as .objs/file.o regardless of where the
.c file is located and fudging the .c into .o would not be accurate.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>