curler: Preserve right errno in try_connect()
In try_connect() we do a non-blocking connect(2) then use poll(2) to
wait for the connect() to succeed or timeout.

We then use getsockopt(2) to see if there was any error on the socket,
i.e if the connect() failed and why. If it failed, we close the sockfd
and log the fact we then set errno to optval (if the getsockopt() itself
didn't fail) to get the reason for the connect() failure.

However we may get errors from the close(2) so do that first.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>