mtd.c: Fix potential issue in mkdir_p()
GitHub code scanning picked up a potential problem in ac_fs_mkdir_p()
where we could overwrite the last '\0' in mdir[] with a '/'.

We need to make mdir 1 bigger to allow for the fact that we add a
trailing '/' to the end of the directory path each time around the loop,
ready for the next directory component to get added.

Switch to using PATH_MAX for the buffer size and add 1 to it, we can
then just check against PATH_MAX for the path length rather than the
size of the mdir buffer.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>