endpoints: Add a wrapper around do_ep()
Previously for each endpoint function we'd make a call to do_ep() like

  return do_ep(IC_CR_CRYSTALLISE, API_VER,
               dsctx, buf, tax_year, (char *)NULL);

There is quite a lot of repetitiveness in all these calls.

They _all_ take API_VER, buf and the NULL terminator.

Create a macro wrapper EP() that takes at a minimum the endpoint and
either dsctx or NULL. Then any extra parameters in the order they are
passed in. So the above would become

  EP(IC_CR_CRYSTALLISE, dsctx, tax_year)

For a GET or DELETE endpoint dsctx would be specified as NULL like
before.

Note: No need for the return or the end ';' as the macro will put them
in.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>