mtd-hmrc-err: Re-work mtd_hmrc_err_map[]
This array is a mapping of MTD error codes to their string
representation.

It is quite a tedious thing to write with some near duplication.

We can improve the situation somewhat by creating a macro to do most of
the work of creating each entry, by simply passing the error code to it
and then let it create the index initialised member.

We keep some members in the old format as they use the full
MTD_HMRC_ERR_ name for the string.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>