Reduce duplication in endpoints[]
Most of the API endpoints have the same .authz and .scope values.

Extend and rename the api_version_map[] array to now be a set of default
values for an endpoint API. Move it into api_endpoints.h where it's
better suited and we can then create a couple of inline helper functions
for getting the right .authz and .scope values.

We check the endpoints array first for the .authz and .scope values, if
they aren't set we grab one or both values from this new array.

This allows to remove most of the .authz and .scope initialisers from
endpoints[]. We leave the OAuth entries as they are slightly different.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>