Revert "curler: Fix formatting of real numbers in the JSON response"
This partly reverts commit 609b24adaaa355a8ffa50e0bafd3a1c07938790d.

It reverts the JSON_REAL_PRECISION(16) part.

So the above commit was made in an attempt to try and fix real numbers
like

  565.56

being shown/written as

  565.55999999999995

and while it helped it didn't fix every case, like the above number.

In the meantime jansson (the JSON library in use) gained a new mechanism
for dealing with such things in the form of David M. Gay's dtoa work.

This seems to resolve the issue, if you leave the default precision at
17.

This patch is in the 2.14.1 release of jansson, so that kind of becomes
a hard requirement for the proper functioning of all this.

Link: <https://github.com/akheron/jansson/commit/8b975abca1055d40637c90b1dc4585af1d7df76c>
Link: <https://github.com/akheron/jansson/releases/tag/v2.14.1>
Signed-off-by: Andrew Clayton <ac@sigsegv.uk>