Flow some extra response headers from HMRC through
When an API is in the process of being deprecated, HMRC will flow the
following response headers through

Deprecation	The deprecation date/time for this endpoint.

Sunset		(optional) The earliest date/time this endpoint will
		become unavailable after deprecation.

Link		Documentation URL for the relevant API.

These will be shown in the main response JSON as either the value as
returned by HMRC or 'null' if they are not set.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>