Allow to log to a file
Allow an application to have the various logging written to a file of
their choice.

This adds a new log_fp structure member to mtd_cfg, which can point to a
FILE * file-pointer of an opened file.

If it's NULL then logging is to stderr/stdout as usual. If it's a valid
FILE * then logs will be sent there instead, except for MTD_LOG_ERROR
logs which are still sent to stderr.

When writing to a file each log messages is preceded by a line like

-- [2025-06-18 19:28:20]

This can be useful to help distinguish new messages from old messages.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>