Report multiple errors from mtd_hmrc_get_error()
A MTD API endpoint can produce multiple errors, e.g.

  "result": {
      "code": "INVALID_REQUEST",
      "message": "Invalid request",
      "errors": [
          {
              "code": "FORMAT_BUSINESS_ID",
              "message": "The Business ID format is invalid"
          },
          {
              "code": "FORMAT_TYPE_OF_BUSINESS",
              "message": "The provided type of business is invalid"
          }
      ]
  }

Previously we'd just report the error as INVALID_REQUEST or maybe
HMRC_MTD_ERR_MULTIPLE_ERRORS.

However it would be much more useful to report _all_ the errors from the
errors array.

This commit modifies the mtd_hmrc_get_error() function to return a
linked list of structures, one for each error.

This is done via a new structure

  struct mtd_hmrc_err {
          enum mtd_hmrc_error error;
          const char *code;
          const char *msg;
          struct mtd_hmrc_err *next;
  }

The next member points to the next structure or it's NULL.

The code and msg members are the code and message strings from the
errors array objects.

This should be freed by mtd_hmrc_free_error().

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>